refactor: extract flashloan calculations and adjust logic to their own modules
