+++ packages/domain/src/flashloans.ts
@@ -1,0 +1,5 @@
+/** For example, flashloaning DAI to open an ETH/USDC position on AAVE */
+export const transientCollateralFlashloan = () => {}
+
+/** For example, flashloaning USDC to open an ETH/USDC position on Ajna */
+export const debtToCollateralSwapFlashloan = () => {}
